requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
joblib==1.5.3
numpy==2.4.2
pandas==3.0.0
python-dateutil==2.9.0.post0
python-telegram-bot==20.7
requests==2.32.5
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
